Encode text metadata to bytes before adding to PNG image
Description:
This change ensures that any text metadata extracted from the source PNG image is encoded as bytes (UTF-8) before being added to the target PNG image. This is done to prevent an error that occurs if the metadata is in string format, as the PNG specification requires that text metadata is stored as bytes.